Load 3.3.482 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1031 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/actions.c
@@ -1374,7 +1374,7 @@ e_add_action (svalue_t *func, svalue_t *cmd, int flag)
         }
         else if (flag < AA_VERB)
         {
-            if (-flag >= mstrsize(p->verb))
+            if ((size_t)(-flag) >= mstrsize(p->verb))
             {
                 free_action_sent(p);
                 error("Bad arg 3 to add_action(): value %ld larger than verb '%s'.\n"
