+++ src/mstrings.c
@@ -1382,7 +1382,7 @@ mstring_extract (const string_t *str, size_t start, long end MTRACE_DECL)
     if (end < 0)
         end = (long)len-1;
 
-    if (end >= len)
+    if (end >= (long)len)
     {
         error("(mstring_extract) end %ld >= len %lu\n"
              , end, (unsigned long) len);
